endpoint login and add some condition on todolist controller

+++ app/services/todo_service/destroy_todo_service.rb
@@ -1,12 +1,18 @@
 class TodoService::DestroyTodoService
 
-  def initialize(id)
+  def initialize(id, current_user)
     @id = id
+    @current_user = current_user
   end
 
   def call
-    todo = Todo.find_by id: @id
-    todo.destroy
+    todo = Todo.where(user_id: @current_user.id).find_by id: @id
+    if todo.nil?
+      return todo
+    else
+      todo.destroy
+      todo
+    end
   end
 
 end
